Initial reorg proposition
