+++ crates/prefab/src/prefab/component/camera.rs
