+++ crates/prefab/src/prefab/component/material.rs
