+++ crates/prefab/src/prefab/component/path.rs
