+++ crates/prefab/src/prefab/component/shape.rs
