+++ examples/spawn_prefab.rs
@@ -1,6 +1,6 @@
 use bevy::prelude::*;
 use bevy_panorbit_camera::PanOrbitCameraPlugin;
-use space_editor::prelude::*;
+use prefab::prefab::{load::PrefabBundle, PrefabPlugin};
 
 fn main() {
     App::default()
