--- src/prefab/component/light.rs
@@ -1,1 +1,0 @@
-
